requirements.txt
Django>=4.2.0
Pillow>=10.0.0
pytesseract>=0.3.10
opencv-python>=4.8.0
numpy>=1.24.0 